BUG: Fixed rounding for f64 to f32 conversion, we should not be rounding to an int!  BUG: Fixed off by .5 problem for 1D and 3D texture interpolation.
Former-commit-id: 6f67286a0f9cd215b3db126ef7b90aa1c6423a86

+++ ocelot/ocelot/ir/implementation/PTXOperand.cpp
@@ -343,7 +343,7 @@ static std::ostream & write(std::ostream &stream, float value) {
 */
 static std::ostream & write(std::ostream &stream, double value) {
 	union {
-		unsigned int imm_uint;
+		long long unsigned int imm_uint;
 		double value;
 	} double_union;
 	double_union.value = value;
